Change source to git
The latest git version greatly improve the experience than the older
release (from 2020-Nov-01). This also allows the build to be compiled
using the latest platforms without adding a number of patches in the
manifest for the messon build.

No stability issues or bugs were observed in testing this change.